fix: cast cosine similarity return to Python float in ToolRegistry
_cosine_similarity was annotated -> float but returned np.float32 due to
numpy preserving input dtype. Added float() cast at the return site.

Also adds contract tests verifying the Unit 4 / Unit 6 integration seams
(import resolution, function signatures, type compatibility, configuration
constants, shape compatibility).

+++ backend/src/agent/registry.py
@@ -259,4 +259,4 @@ class ToolRegistry:
 
         # Cosine similarity = dot(a, b) / (norm(a) * norm(b))
         dot_product = np.dot(a_padded, b_padded)
-        return dot_product / (norm_a * norm_b)
+        return float(dot_product / (norm_a * norm_b))
